Fix tabbing
At the moment the new tabDisable option has usability problems:

* In classical mode a `tab` focuses on the toolbar of the next editor,
  while we want to go to the edit area of the next editor
* In inline mode, this problem doesn't exist (because the toolbars are
  not displayed) but `shift-tab` doesn't work (the focus may be in the
  hidden toolbar)

So here is a proposed solution by setting tabindex="-1"
cf. https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/tabindex

Maybe it's not the right way to fix the problem, but it gives you a hint
of what it could be.