css restructuring with themes
WHAT WAS DONE?

The css has been restructured to allow the simplified use of themes through color variables. That way, each theme only needs to overwrite the colors in ":root" or create a new configuration.

WHAT DO YOU NEED TO MODIFY/ADD TO THE CODE?
-------------------------------------------------- ------------------------------------------
- Include a "theme" option. For example:
     const initEditor=suneditor.init({
         "name-of-theme": "light/dark/yellow/blue/...",
     ...

- Add "<body data-se-theme="name-of-theme">"

TO CREATE A NEW THEME
------------------------------------------------
- Duplicate the section "[data-se-theme="light"]" with a new name and change the desired colors.
- As an additional suggestion, you can create themes in separate .js files and include the desired theme file in the application's header, just as you do with language files.

Only this is needed.